Refactor skills and links sections in UserCard component into own components.
