+++ src/components/Home/UserCard/UserCard.css
@@ -79,72 +79,3 @@
     font-size: 0.8rem;
     border-bottom: 1px solid black;
 }
-
-.card-link-container {
-    width: 100%;
-    height: 4rem;
-
-    border-bottom: 1px solid black;
-    
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-}
-
-.card-link {
-    width: 100%;
-    height: 1.1rem;
-    padding: 0rem 0.4rem;
-    
-    overflow: hidden;
-    font-weight: bold;
-    text-decoration: none;
-
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-.card-link> svg {
-    font-size: 1.1rem;
-}
-
-.card-skill-container {
-    width: 100%;
-    height: 1.5rem;
-
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    gap: 0.5rem;
-}
-
-.card-skill {
-    position: relative;
-    width: auto;
-    height: min-content;
-    max-height: 1.3rem;
-
-    overflow: hidden;
-    border-radius: 0.5rem;    
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.142);
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.card-skill > img {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-}
-
-.card-skill-overlay {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-
-    background-color: transparent;
-    box-shadow: inset 0 -4px 8px #585858ee;
-    border-radius: 0.5rem;    
-}
