+++ src/components/Home/UserCard/UserCard.jsx
@@ -1,8 +1,7 @@
 import PropTypes from "prop-types";
-import { FaHackerrank, FaGithub, FaLinkedin, FaInstagram, FaFacebook, FaDiscord, FaLink } from "react-icons/fa";
-import { CgPokemon } from "react-icons/cg";
 import "./UserCard.css";
-import markdownBadges from "../../../assets/markdown-badges.json"
+import UserSkills from "./UserSkills/UserSkills";
+import UserLinks from "./UserLinks/UserLinks";
 
 const UserCard = ({ user }) => {
   return (
@@ -30,76 +29,6 @@ const UserCard = ({ user }) => {
   );
 };
 
-
-const UserLinks = ({ links }) => {
-  let linkComponents = [];
-  const iconReference = {
-    "linkedin": <FaLinkedin />,
-    "github": <FaGithub />,
-    "hackerrank": <FaHackerrank />,
-    "instagram": <FaInstagram />,
-    "facebook": <FaFacebook />,
-    "discord": <FaDiscord />,
-    "pokemon": <CgPokemon />
-  }
-
-  for (let i = 0; i < 3; i++) {
-    const linkName = Object.keys(links)[i]
-    let linkLogo = <FaLink />
-    if (Object.keys(iconReference).includes(linkName.toLowerCase())) {
-      linkLogo = iconReference[linkName.toLowerCase()];
-    }
-    linkComponents[i] = (
-      <a
-        className="card-link"
-        href={links[linkName]}
-        key={`link${i}`}
-      >
-        {linkLogo}
-        {linkName}
-      </a>
-    )
-  }
-
-  return (
-    <div className="card-link-container">
-      {linkComponents}
-    </div>
-  )
-}
-
-
-const UserSkills = ({ skills }) => {
-  let skillComponents = [];
-  for (let i = 0; i < 3; i++) {
-    const skillName = skills[i]
-    let skillBadge = (
-      <div 
-        className="card-skill" 
-        key={`user skill${i}`}>
-        {skillName}
-      </div>
-    )
-    if (Object.keys(markdownBadges).includes(skillName)) {
-      skillBadge = (
-        <div
-          className="card-skill"
-          key={`user skill${i}`}>
-          <div className="card-skill-overlay"></div>
-          <img src={markdownBadges[skillName]} />
-        </div>
-      )
-    }
-    skillComponents[i] = skillBadge 
-  }
-
-  return (
-  <div className="card-skill-container">
-    {skillComponents}
-  </div>
-  )
-}
-
 UserCard.propTypes = {
   user: PropTypes.shape({
     name: PropTypes.string,
@@ -111,13 +40,5 @@ UserCard.propTypes = {
     skills: PropTypes.arrayOf(PropTypes.string),
   }),
 };
-
-UserLinks.propTypes = {
-  links: PropTypes.object,
-}
-
-UserSkills.propTypes = {
-  skills: PropTypes.arrayOf(PropTypes.string)
-}
 
 export default UserCard;
