+++ src/components/Home/UserCard/UserLinks/UserLinks.css
@@ -1,0 +1,27 @@
+.card-link-container {
+    width: 100%;
+    height: 4rem;
+
+    border-bottom: 1px solid black;
+    
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+}
+
+.card-link {
+    width: 100%;
+    height: 1.1rem;
+    padding: 0rem 0.4rem;
+    
+    overflow: hidden;
+    font-weight: bold;
+    text-decoration: none;
+
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+.card-link> svg {
+    font-size: 1.1rem;
+}
