+++ src/components/Home/UserCard/UserSkills/UserSkills.css
@@ -1,0 +1,40 @@
+.card-skill-container {
+    width: 100%;
+    height: 1.5rem;
+
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.card-skill {
+    position: relative;
+    width: auto;
+    height: min-content;
+    max-height: 1.3rem;
+
+    overflow: hidden;
+    border-radius: 0.5rem;    
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.142);
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.card-skill > img {
+    width: 100%;
+    height: 100%;
+    object-fit: contain;
+}
+
+.card-skill-overlay {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+
+    background-color: transparent;
+    box-shadow: inset 0 -4px 8px #585858ee;
+    border-radius: 0.5rem;    
+}
